Actualizacion JAMedia 16 c++

+++ JAMedia/c++/PlayerList/JTreeView.cpp
@@ -37,8 +37,7 @@ void JTreeView::accion_menu(Glib::ustring text, Gtk::TreePath path){
     JAMedia *top = dynamic_cast<JAMedia*> (this->get_toplevel());
 
     if (text == "Quitar"){
-        if (selected == process){next_track();}
-        listore->erase(iter);
+        if (selected == process){next_track();}listore->erase(iter);
         Gtk::TreeModel::iterator iter = sel->get_selected();
         if (not iter){top->init();}}
 
@@ -54,41 +53,47 @@ void JTreeView::accion_menu(Glib::ustring text, Gtk::TreePath path){
 
         switch (result){
             case Gtk::RESPONSE_OK:{
-                if (selected == process){next_track();}
-                delete_file(process);
+                if (selected == process){next_track();}delete_file(process);
                 listore->erase(iter);
                 Gtk::TreeModel::iterator iter3 = sel->get_selected();
-                if (not iter3){top->init();}
-                break;}
+                if (not iter3){top->init();}break;}
         delete label; dialog.hide();}}
 
-    else if (text == "Guardar"){Glib::ustring destino = save_file(process);}
-    else if (text == "Mover"){Glib::ustring destino = save_file(process);
-        if (destino != ""){delete_file(process);row[cols.path] = destino;}} //FIXME: antes lo quitabamos de la lista
+    else if (text == "Copiar"){save_file(process, text);}
+    else if (text == "Mover"){if (selected == process){next_track();}
+        save_file(process, text);listore->erase(iter);
+        Gtk::TreeModel::iterator iter3 = sel->get_selected();
+        if (not iter3){top->init();}}
+
+    else if (text == "Subtitulos"){top->load_sub();}
     else if (text == "Grabar"){}
         //FIXME: Grabar/Convertir/Extraer en el directorio seleccionado
 }
 
 void JTreeView::delete_file(Glib::ustring process){
-    if (std::ifstream(process.c_str())){
-        if (std::remove(process.c_str()) != 0){
-            std::cout << "ERROR al intentar borrar: " << process << std::endl;}
-        else {std::cout << "ARCHIVO ELIMINADO: " << process << std::endl; }}}
+    Glib::RefPtr<Gio::File> giofile =
+        Gio::File::create_for_parse_name(process);
+    if (giofile->trash()){
+        std::cout << "ARCHIVO ELIMINADO: " << process << std::endl;}
+    else{std::cout << "ERROR al intentar borrar: " << process << std::endl;}}
+    //if (std::ifstream(process.c_str())){
+    //    if (std::remove(process.c_str()) != 0){
+    //        std::cout << "ERROR al intentar borrar: " << process << std::endl;}
+    //    else {std::cout << "ARCHIVO ELIMINADO: " << process << std::endl; }}}
     //FIXME: Si falla borrar, es un track en un archivo de lista de reproduccion (pls, meu, etc)
 
-Glib::ustring JTreeView::save_file(Glib::ustring process){
+void JTreeView::save_file(Glib::ustring process, Glib::ustring accion){
     JAMedia *top = dynamic_cast<JAMedia*> (this->get_toplevel());
-    Gtk::FileChooserDialog dialog(*top, "Guardar Archivo",
+    Gtk::FileChooserDialog dialog(*top, accion + " Archivo",
     Gtk::FILE_CHOOSER_ACTION_SAVE);
-    dialog.add_button("Guardar", Gtk::RESPONSE_OK);
+    dialog.add_button(accion, Gtk::RESPONSE_OK);
     dialog.add_button("Cancelar", Gtk::RESPONSE_CANCEL);
     dialog.set_do_overwrite_confirmation(true);
     dialog.set_border_width(15); dialog.set_filename(process);
-    Glib::RefPtr<Gio::File> gioFile; Glib::RefPtr<Gio::FileInfo> info;
-    gioFile = Gio::File::create_for_path(process);
-    info = gioFile->query_info();
+    Glib::RefPtr<Gio::File> giofile =
+        Gio::File::create_for_parse_name(process);
+    Glib::RefPtr<Gio::FileInfo> info;info = giofile->query_info();
     Glib::ustring tipo = info->get_content_type();
-    std::cout << tipo << std::endl;
     Glib::RefPtr<Gtk::FileFilter> filter = Gtk::FileFilter::create();
     size_t found = tipo.find("audio");
     if (found!=std::string::npos){filter = Gtk::FileFilter::create();
@@ -98,19 +103,27 @@ Glib::ustring JTreeView::save_file(Glib::ustring process){
     if (found!=std::string::npos){filter = Gtk::FileFilter::create();
         filter->set_name("Videos"); filter->add_mime_type("video/*");
         dialog.add_filter(filter);}
-    found = tipo.find("vnd.rn-realmedia");
-    if (found!=std::string::npos){filter = Gtk::FileFilter::create();
-        filter->set_name("Videos");
-        filter->add_mime_type("application/vnd.rn-realmedia/*");
-        dialog.add_filter(filter);}
     Glib::ustring destino = ""; int result = dialog.run();
     switch (result){case Gtk::RESPONSE_OK:
     destino = dialog.get_filename().c_str(); break;} dialog.hide();
-    if (destino != ""){std::FILE *f1 = std::fopen(process.c_str(), "rb");
-        std::FILE *f2 = std::fopen(destino.c_str(), "wb");
-        if (f1 != NULL and f2 != NULL){unsigned int c;
-            while(!std::feof(f1)){c = std::getc(f1); std::fputc(c, f2);}}
-        std::fclose(f1); std::fclose(f2); return destino;}}
+    if (destino != ""){
+        Glib::RefPtr<Gio::File> giodest =
+            Gio::File::create_for_parse_name(destino);
+        if (accion == "Copiar"){
+            if (not giofile->copy(giodest)){
+                std::cout << "ERROR al Copiar: " << process;
+                std::cout << " en: " << destino << std::endl;}}
+        else if (accion == "Mover"){
+            if (not giofile->move(giodest)){
+                std::cout << "ERROR al Mover: " << process;
+                std::cout << " en: " << destino << std::endl;}}
+        else{std::cout << accion << std::endl;}
+        }}
+    //if (destino != ""){std::FILE *f1 = std::fopen(process.c_str(), "rb");
+    //    std::FILE *f2 = std::fopen(destino.c_str(), "wb");
+    //    if (f1 != NULL and f2 != NULL){unsigned int c;
+    //        while(!std::feof(f1)){c = std::getc(f1); std::fputc(c, f2);}}
+    //    std::fclose(f1); std::fclose(f2);return destino;}}
 
 bool JTreeView::on_button_press_event(GdkEventButton *event){
     //click izq selecciona, der menu contextual.
@@ -120,9 +133,19 @@ bool JTreeView::on_button_press_event(GdkEventButton *event){
     Gtk::TreeModel::iterator iter = listore->get_iter(path);
     if ((event->type == GDK_BUTTON_PRESS) && (event->button == 1)){
         sel->select(iter);
-        row_activated(listore->get_path(iter), get_column(0));return true;}
+        row_activated(listore->get_path(iter), get_column(0));
+        scroll_to_row(listore->get_path(iter)); return true;}
     else if ((event->type == GDK_BUTTON_PRESS) && (event->button == 3)){
-        JMenu *menu = new JMenu(path);menu->attach_to_widget(*this);
+        Gtk::TreeModel::Row row = *listore->get_iter(path);
+        Glib::RefPtr<Gio::File> giofile =
+            Gio::File::create_for_parse_name(row[cols.path]);
+        bool video = false;
+        if(giofile->query_exists()){
+            Glib::ustring tipo = giofile->query_info()->get_content_type();
+            size_t found = tipo.find("video");
+            if (found != std::string::npos){video = true;}}
+        JMenu *menu = new JMenu(path, giofile->query_exists(), video);
+        menu->attach_to_widget(*this);
         menu->popup(event->button, event->time);return true;}
     else {return false;}}
 
@@ -131,30 +154,30 @@ void JTreeView::select_begin(){
     //if(row) sel->select(row);
     Gtk::TreeModel::iterator iter = listore->children().begin();
     if (iter){sel->select(iter);
-        row_activated(listore->get_path(iter), get_column(0));}}
+        row_activated(listore->get_path(iter), get_column(0));
+        scroll_to_row(listore->get_path(iter));}}
 
 void JTreeView::select_end(){
     Gtk::TreeModel::iterator iter = listore->children().end();
     Gtk::TreeModel::Row row = *(--iter);
     if (row){sel->select(iter);
-        row_activated(listore->get_path(iter), get_column(0));}}
+        row_activated(listore->get_path(iter), get_column(0));
+        scroll_to_row(listore->get_path(iter));}}
 
 void JTreeView::previous_track(){
     Gtk::TreeModel::iterator iter = sel->get_selected();
-    if (iter){
-        Gtk::TreeModel::Row row = *(--iter);
-        if (row){
-            sel->select(iter);
-            row_activated(listore->get_path(iter), get_column(0));}
+    if (iter){Gtk::TreeModel::Row row = *(--iter);
+        if (row){sel->select(iter);
+            row_activated(listore->get_path(iter), get_column(0));
+            scroll_to_row(listore->get_path(iter));}
         else{select_end();}}}
 
 void JTreeView::next_track(){
     Gtk::TreeModel::iterator iter = sel->get_selected();
-    if (iter){
-        Gtk::TreeModel::Row row = *(++iter);
-        if (row){
-            sel->select(iter);
-            row_activated(listore->get_path(iter), get_column(0));}
+    if (iter){Gtk::TreeModel::Row row = *(++iter);
+        if (row){sel->select(iter);
+            row_activated(listore->get_path(iter), get_column(0));
+            scroll_to_row(listore->get_path(iter));}
         else{select_begin();}}}
 
 //bool JTreeView::select_function(const Glib::RefPtr<Gtk::TreeModel>& model,
@@ -174,35 +197,64 @@ void JTreeView::row_activated(const Gtk::TreePath &path,
     Gtk::TreeModel::iterator iter = listore->get_iter(path);
     if (iter){Gtk::TreeModel::Row row = *iter;
         JAMedia *top = dynamic_cast<JAMedia*> (this->get_toplevel());
-        top->load_file(row[cols.path]);}} //FIXME: scroll_to_cell(&path, &col);
+        top->load_file(row[cols.path]);}}
 
 void JTreeView::clear_list(){listore->clear();}
 
 void JTreeView::open_files(std::vector<std::basic_string<char> > lista){
-    //Agrega pistas a la lista
-    Gtk::TreeModel::Row row; Glib::RefPtr<Gdk::Pixbuf> pixbuf;
-    Glib::ustring icon = "./Iconos/sonido.svg";
-    Glib::RefPtr<Gio::File> gioFile; Glib::RefPtr<Gio::FileInfo> info;
-    Glib::ustring tipo;
+    //Recibe una lista de archivos que pueden ser archivos de audio y/o video
+    //o un archivo de lista de reproducción en formato pls, m3u, etc...
     for(int i = 0; i < lista.size(); i++){
         std::basic_string<char> path = lista[i];
         Glib::ustring name = basename(path.c_str());
-        Glib::ustring rpath = path.c_str();
-        gioFile = Gio::File::create_for_path(rpath);
-        info = gioFile->query_info();
-        tipo = info->get_content_type(); row = *(listore->append());
-        try{size_t found = tipo.find("audio");
-            if (found!=std::string::npos){icon = "./Iconos/sonido.svg";}
-            else {found = tipo.find("video");
-                //FIXME: Agregar "application/vnd.rn-realmedia/*"
-                if (found!=std::string::npos){icon = "./Iconos/video.svg";}
-                else{icon = "./Iconos/sonido.svg";
-                    std::cout << "FIXME: Nuevo tipo de archivo ";
-                    std::cout << (found!=std::string::npos);
-                    std::cout << " " << tipo << std::endl;}}
+        size_t ext = name.find(".pls"); //FIXME: m3u, json, JAMedia
+        if (ext != std::string::npos){pls_read(path);}
+        else{append_file_track(path);}}}
+
+void JTreeView::pls_read(std::basic_string<char> path){
+    std::basic_string<char> line;
+    std::ifstream archivo;
+    archivo.open(path.c_str());
+    if (archivo.is_open()){
+        Glib::ustring title;
+        Glib::ustring track;
+        while(std::getline(archivo, line)){
+            size_t pos_tag = line.find("File");
+            if (pos_tag != std::string::npos){
+                track = line.substr(line.find("=")+1);}
+            else{pos_tag = line.find("Title");
+                if (pos_tag != std::string::npos){
+                    title = line.substr(line.find("=")+1);}}
+            if (not title.empty() and not track.empty()){
+                //FIXME: quitar espacios y tabulaciones, comillas, etc.
+                Glib::RefPtr<Gio::File> giofile =
+                    Gio::File::create_for_parse_name(path);
+                if (giofile->query_exists() or gst_uri_is_valid(path.c_str())){
+                    Gtk::TreeModel::Row row = *(listore->append());
+                    Glib::ustring icon = "./Iconos/sonido.svg";
+                    Glib::RefPtr<Gdk::Pixbuf> pixbuf =
+                        Gdk::Pixbuf::create_from_file(icon);
+                    pixbuf = pixbuf->scale_simple(24, 24, Gdk::INTERP_BILINEAR);
+                    row[cols.pix] = pixbuf;row[cols.name] = title;
+                    row[cols.path] = track;}track.clear(); title.clear();}
+            }archivo.close();}}
+
+void JTreeView::append_file_track(std::basic_string<char> path){
+    //recibe el path a un archivo.
+    Gtk::TreeModel::Row row = *(listore->append());
+    Glib::ustring icon = "./Iconos/sonido.svg";
+    Glib::RefPtr<Gdk::Pixbuf> pixbuf = Gdk::Pixbuf::create_from_file(icon);
+    pixbuf = pixbuf->scale_simple(24, 24, Gdk::INTERP_BILINEAR);
+    Glib::RefPtr<Gio::File> giofile = Gio::File::create_for_parse_name(path);
+    //Glib::RefPtr<Gio::FileInfo> info;
+    Glib::ustring tipo = giofile->query_info()->get_content_type();
+    Glib::ustring name = basename(path.c_str());
+    try{//FIXME: Agregar "application/vnd.rn-realmedia/*"
+        size_t found = tipo.find("video");
+        if (found != std::string::npos){icon = "./Iconos/video.svg";
             pixbuf = Gdk::Pixbuf::create_from_file(icon);
             pixbuf = pixbuf->scale_simple(24, 24, Gdk::INTERP_BILINEAR);
-            row[cols.pix] = pixbuf;}
-        catch(const Glib::FileError& e){std::cout << e.what() << std::endl;
-            row[cols.pix] = pixbuf;}
-        row[cols.name] = name; row[cols.path] = rpath;}}
+            row[cols.pix] = pixbuf;}}
+    catch(const Glib::FileError& e){std::cout << e.what() << std::endl;
+        row[cols.pix] = pixbuf;}
+    row[cols.name] = name;row[cols.path] = path.c_str();}
